# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationSummary": {
        "properties": {
          "booking_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "counterpart_name": {
            "title": "Counterpart Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "last_message_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message Preview"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "request_status": {
            "title": "Request Status",
            "type": "string"
          },
          "request_title": {
            "title": "Request Title",
            "type": "string"
          },
          "service_name": {
            "title": "Service Name",
            "type": "string"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "request_id",
          "service_name",
          "request_title",
          "counterpart_name",
          "request_status",
          "booking_id",
          "last_message_at",
          "last_message_preview",
          "unread_count"
        ],
        "title": "ConversationSummary",
        "type": "object"
      },
      "ConversationView": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationSummary"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageView"
            },
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "conversation",
          "messages"
        ],
        "title": "ConversationView",
        "type": "object"
      },
...
      "EmailDeliveryStatus": {
        "enum": [
          "PENDING",
          "SENT",
          "SKIPPED",
          "FAILED"
        ],
        "title": "EmailDeliveryStatus",
        "type": "string"
      },
...
      "FavoriteProviderView": {
        "properties": {
          "business_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_identity_verified": {
            "title": "Is Identity Verified",
            "type": "boolean"
          },
          "provider_slug": {
            "title": "Provider Slug",
            "type": "string"
          },
          "rating_average": {
            "title": "Rating Average",
            "type": "number"
          },
          "rating_count": {
            "title": "Rating Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "provider_slug",
          "display_name",
          "business_name",
          "rating_average",
          "rating_count",
          "is_identity_verified",
          "created_at"
        ],
        "title": "FavoriteProviderView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MessageCreate": {
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "MessageCreate",
        "type": "object"
      },
      "MessageModerationStatus": {
        "enum": [
          "VISIBLE",
          "FLAGGED",
          "HIDDEN"
        ],
        "title": "MessageModerationStatus",
        "type": "string"
      },
      "MessageView": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_mine": {
            "title": "Is Mine",
            "type": "boolean"
          },
          "moderation_status": {
            "$ref": "#/components/schemas/MessageModerationStatus"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "sender_id": {
            "format": "uuid",
            "title": "Sender Id",
            "type": "string"
          },
          "sender_name": {
            "title": "Sender Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sender_id",
          "sender_name",
          "is_mine",
          "body",
          "moderation_status",
          "read_at",
          "created_at"
        ],
        "title": "MessageView",
        "type": "object"
      },
      "NotificationPage": {
        "properties": {
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/NotificationView"
            },
            "title": "Notifications",
            "type": "array"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "notifications",
          "unread_count"
        ],
        "title": "NotificationPage",
        "type": "object"
      },
      "NotificationType": {
        "enum": [
          "REQUEST_RECEIVED",
          "REQUEST_UPDATED",
          "QUOTE_RECEIVED",
          "BOOKING_UPDATED",
          "MESSAGE_RECEIVED",
          "REVIEW_RECEIVED",
          "REVIEW_MODERATED"
        ],
        "title": "NotificationType",
        "type": "string"
      },
      "NotificationView": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email_status": {
            "$ref": "#/components/schemas/EmailDeliveryStatus"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "link_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Path"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "resource_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "link_path",
          "resource_type",
          "resource_id",
          "read_at",
          "email_status",
          "created_at"
        ],
        "title": "NotificationView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SIC API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/client/favorites": {
      "get": {
        "operationId": "favorites_v1_client_favorites_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FavoriteProviderView"
                  },
                  "title": "Response Favorites V1 Client Favorites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Favorites",
        "tags": [
          "client-favorites"
        ]
      }
    },
    "/v1/client/favorites/{provider_slug}": {
      "delete": {
        "operationId": "remove_favorite_v1_client_favorites__provider_slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider_slug",
            "required": true,
            "schema": {
              "title": "Provider Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Favorite",
        "tags": [
          "client-favorites"
        ]
      },
      "put": {
        "operationId": "add_favorite_v1_client_favorites__provider_slug__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider_slug",
            "required": true,
            "schema": {
              "title": "Provider Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteProviderView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Favorite",
        "tags": [
          "client-favorites"
        ]
      }
    },
...
    "/v1/me/conversations": {
      "get": {
        "operationId": "conversations_v1_me_conversations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationSummary"
                  },
                  "title": "Response Conversations V1 Me Conversations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Conversations",
        "tags": [
          "messaging"
        ]
      }
    },
    "/v1/me/conversations/{request_id}": {
      "get": {
        "operationId": "conversation_v1_me_conversations__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Conversation",
        "tags": [
          "messaging"
        ]
      }
    },
    "/v1/me/conversations/{request_id}/messages": {
      "post": {
        "operationId": "send_message_v1_me_conversations__request_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "messaging"
        ]
      }
    },
    "/v1/me/notifications": {
      "get": {
        "operationId": "notifications_v1_me_notifications_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/me/notifications/read-all": {
      "post": {
        "operationId": "mark_all_notifications_read_v1_me_notifications_read_all_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark All Notifications Read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/me/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_notification_read_v1_me_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Notification Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
  }
}